+++ NEWS.md
@@ -1,5 +1,7 @@
 # lipdR 0.5.2
 
+* make `readLipd()` give an error if a local file doesn't exist.
+
 # lipdR 0.5.1
 
 * Make `readLipd()` fail better when downloading from lipdverse or other urls